Add sprints' Start Date
Add in the structure, search criteria and columns.

Fill with the values provided by the REST API.

I only care about startDate here to put in place
the test mechanism (xpath) and the new classes.

The integration tests are update to xpath to locate
the elements to remove a bit of coupling between
the tests and the implementation. For instance we
only really care in the tests if start_date formElement
is present, it's actual placement (in which container)
is not really important here.

Part of story #19242 create Jira project with Agile configuration

Change-Id: I23cc4304daa201f2965bc74588b56bc370b9a237