Add the sprint state select box
In order to have a compeling test for references the whole
structure is done in one commit (formElement, report, criterion
table column and semantic).

The actual value will be managed in a dedicated commit.

This is mandatory to have open & closed milestones in the
planning view.

Part of story #19242 create Jira project with Agile configuration

Change-Id: I151ad12aee5174e28dbc41aea6b0b1187ae5a467