Remove the modal v2 from legacy script of Tracker
Part of story #18336 add support of markdown

This first step is needed to be able to use the brand new rich text editor in the modal v2.

There is no functional change expected.
In the Web Console, you should see the file modal-v2-***.js loaded.

This patch also help for the request #19306

Change-Id: If56e2ab9068c8658e4f4e4965e1415182e0c3ec6