request #19282 : One test should always take focus in tests list
This makes sure a test in tests list, and one only, always has `tabindex="0"`.
Therefore user should always be able to enter the tests list.

We need to set tabindex upon page load and filtering ;
Using `execution-list-filter` as a workaround to trigger this function allows to do that.

However, filtering out and straight back in an active test tab will leave `tabindex=0` on both first and active test tabs.

Change-Id: I911faab4269d0ba918c016d02a07da6a271f8ed5