Fatal error when summary service does not exist
If the summary service is not present in the project, then we should not
have a fatal error when we want to manage services in project
administration.

This can be reproduced with XML import with `<services/>`.

While we probably should not allow to create a project without any
services, and/or without summary, at least we have to remove the fatal
error for existing imports.

Fix request #19307: Project import can create unusable project

Change-Id: I9ff36cf5f72bb4cfcffacc8b1e348d28a34a2c5a