Import worklog comment
This is part of story #19253 import worklog as timetracking

If there is a comment added with a worklog, this comment is imported.
Only the raw text parts of this comment is added.

If the user is not found in mapping, an information about who added the
worklog is added in comment (as done for issue comments).

Change-Id: Iee0d69fa08661340c6190ba8a1aa7087ae6d38b5