request #19320 Be able to select an artifact parent at artifact update
Given a project with a defined tracker hierarchy, if I update an artifact in a
tracker that is defined as child of another tracker, and given that this artifact
does not have an artifact parent yet, then the parent selector in artifact link
field must be displayed at artifact update in the tracker artifact UI.

This patch does not allow creation of a new parent at artifact update.
It will be handled in a dedicated commit.

Change-Id: If817c2f2525f9f7306477e025c05dc5004ad2370