Refacto: UGroups that can see a private comment must be in the comment object
This is part of story #19275 restrict who can see comments

No functional changes.
- Import XML with comments with private ugroups must save ugroups in
  plugin_tracker_private_comment_permission table
- If you haven't rights to see a comment, in API Rest
  artifact/$id/changeset you should not see it

Instead of getting ugroups that can see a private comment in PermissionChecker.php,
now, they are injected in Comment object.

This is needed for the next patch: to avoid making the same call to database to
get these ugroups (once in permission checker, once in comment xml exporter).

Change-Id: I19e1fe56a28d3aef0b5518f68d5920c38c3bc5ac