request #19336: Remove user lab mode
The lab mode in the way it is currently designed has not been used
these past years and can be a source of questions for users as they
have no way to know what's behind it.

Scrum v2 (scrum mono milestone) is now hidden behind a feature flag
instead of user lab mode.

This brings a new way to manage feature flags, so that developers have
better ways to identify existing feature flags. For now only scrum v2
and list picker usage in trackers are covered by this proposal.

Change-Id: I88e54b7fe77a4bed948a808dec5b0d2ec8625e4a