Import base of AD planning configuration
This is part of story #19242 create Jira project with Agile configuration

When importing a Jira project, the agiledashboard service is now
activated.

If the Jira project has at least one board, then a Scrum planning is
created in the Tuleap AD service. All non subtask and non Epic issueTypes are
set as backlog trackers of this planning. Sprint tracker created to import
sprint data from Jira is set as the planning tracker.

Given that we cannot find easily the Epic issueType [1], then a new
option is provided to declare what is the Epic issueType of the project:

--jira-epic-issue-type=EpicEN

If not provided, the value Epic will be used by default.

[1] https://community.developer.atlassian.com/t/how-to-determine-is-issue-via-rest-api-an-epic-with-non-en-en-locale/33209/2

Change-Id: I70a899e1d920a3e97a22de20bdc550dc5a171cc9