request #19346 Frs rest routes return error if user is not in file info
The value user on file info can be null on old package.

This condition is not respected on file info representation. If we try to get this file or the related package with rest api, it throw an internal server error.

To test it, you can update the user_id value to null on frs_file table and try to get the package or file with rest routes.

Change-Id: I35bfea11d71c47ba80879ca189ecafa9f35982e4