Use List-Id header for artifacts notifications
Whenever the summary of an artifact changes (for example after a
typo) or `[assigned to]` appears in the email notification subject,
the email client might be lost and thinks that it is a new thread.
In order to help it to better organize the artifact email
notifications, we can borrow the
[List-Id header](https://tools.ietf.org/html/rfc2919).

In the mail sent after an artifact update, we should have a List-Id:
header.

Closes request #19354: Use List-Id header for artifacts notifications

Change-Id: Ib1696c2913f2ba89aecca724b37aaa4c0f8088dd