Widget targets wrong chart after project creation
Given a project A with an tracker pie chart AA

And a project B with an tracker bar chart BB

And two widgets on the dashboard of project B:

* One targeting the pie chart AA (another project)
* One targeting the bar chart BB (same project)

When user creates a new project C based on project B

Then the dashboard of C should have two widgets:

* One targeting the pie chart AA (another project)
* One targeting the bar chart CC which is a copy of BB
and belongs to project C.

Note: since there isn't any artifacts in the new project yet,
the chart appears empty. And since it does not have any empty
state, the widget looks weird. However it is not the aim of
this patch to display an empty state for empty chart.

Closes request #19360: Widget targets wrong chart after project creation

Change-Id: I0366c0a35195760e588690a5d298bd1ec524ee0e