Team user stories should not be planable in PI in order to be able to link feature and PI
partf of story #18408 plan Features in Program Increment

How to test:
Be sure that the only thing plannable in PI is feature tracker
Create a PI, add a feature inside
 => in mirrored release, the user story of feature is added

Why:
Previously, we did not correctly understood AC, and first implementation
of display content was incorrect.

Configuration have not been updated between the correction of first
implementation, and I had leftovers in the plannable tracker list (my PI
had team user stories in plannable elements).

This link is meanignless and should be removed.

Change-Id: I38eae8b8e143fe55c514b3ab2a5f91785dc514b9