Cardwall configuration is now imported
This is part of story #19242 create Jira project with Agile configuration

Cardwall is now enabled in "Sprint Plan" with columns found in the Jira board
configuration. No mapping is needed because in Jira, one column is equal
to one status, so default mapping based on status names in Tuleap cardwall is enough.

No XML helper has been done yet for XML cardwall build. It will be done
in a dedicated commit.

Change-Id: Ifcd36c1e9a88a29bd8e76c4a722c49c333ce6eaf