Import Story Points field as initial effort
There is no reliable ways to find all the fields an issue
type can have[1] and it's un-clear whether Story Points
are limited to "Story" issue types.

So the strategy for Story Points is the following:
- The Board Configuration advertise which field is used to
  hold Story Points
- All trackers (except sub-tasks) are created with a Story
  Points field "in case of". At least, in Tuleap, as they
  are in the Agile Dashboard configuration, they can have
  story points.

The semantic stuff will be done in a dedicated commit.

Note: Tracker_FormElementFactory was modified because I
first thought to make the field computed but it's not
relevant here. However, the modifications made here
was worth to keep IMHO.

[1] https://community.developer.atlassian.com/t/rest-api-for-fieldlayoutscheme-and-issuetypescreenscheme/37714/2

Part of story #19242 create Jira project with Agile configuration

Change-Id: Ibf7559a00135f29edee8ef8dd7e5b8b05cc2cfb5