Tracker report comment filter takes into account private permission
This is part of story #19275 restrict who can see comments

All comments user can see (private commments and non private comments)
are now taken into account in the comment report criteria (in both UI
and TQL).

To ease the development, user and tracker objects are injected directly
in the FromWhere object. To will be removed in a refactoring and both
will be injected in the whole advanced query call stack.

Change-Id: Idb8b7cbef0891cf66a099672d1e609fd7eb0f8ce