request #19373 Separate Cypress and Jest TypeScript globals
How to test:
- No functional change
- CI for Jest tests is happy

Notes:
In order to stop mixing Cypress and Jest global definitions, this change
removes the automatic inclusion of all global types. We must now specify
which "lib" we need (e.g. "ES2015" for Promise global) and which global
"types" to include (e.g. "jest" for "expect" global).

@types/node is needed in AgileDashboard and ProgramManagement because
tests use `setImmediate`. It is also needed in TestManagement because
`process.env` is used in the relative date custom element. Otherwise,
only "jest" is added.

I do not precisely understand why, but it seems Jest will cause a
compilation of typescript files from unrelated plugins when running
tests. For example, Jest could fail a test in Timetracking plugin
because of some false error in AgileDashboard or Document plugin. I
suspect it is because they pull @tuleap/tlp-fetch, but I'm not really
sure. Anyway, adding a tsconfig at their root seems to fix this problem
and is harmless otherwise. It will better separate the plugins into
tsconfig projects.

Change-Id: Ia645ef7ac376c977bbc038a95cf3043429ad5fd9