request #19382: Display a big "Unsupported browser" at the top of the pages when IE is used
With the work done in request #19287, we are not able to open the
browser deprecation modals on most of the pages when IE is used and a
lot of things might start to break.

Even if we cannot display the modal anymore, with this contribution we
try to inform the users they should use another browser.

Change-Id: I03cbe2ad35a37c70ff8d279e40b5b1cd6144b200