Dedicated icon for burnup field
When you edit a BurnUp field, or when you want to add one, the icons
should reflect that we are on a burnup and not a burndown.

Same style as burndown icon is applied to not ruin the fengshui.

Note: Images have been optimized with [crunch](https://github.com/chrissimpkins/Crunch)

Fix request #19381: BurnUp field icon shows a burndown

Change-Id: Iacde65b74b728e956f035b6399b4962b18a6e4d3