request #19356 : Skip to main content link in campaign page should not bring user back to campaigns list page
The `Skip to main link` is a link to an anchor preceding the main content, thus using the hash property in the URL.
However, TestManagement already uses this property (example: `testmanagement/?group_id=101#!/campaigns/32`).

This patch hijacks the `skip to main content` link it TTM so that it triggers the focus method on the main content instead.

This problem might occur in other pages using the hash property of the URL.
However, rather than changing the overall way of how the `skip to main content` link works, we change it where it is necessary.

Change-Id: Ic796fb5ba6dc9f28b45d1da09893243c39ae53c3