Better handling of multi trackers import
This is part of request #19188 Check fields permissions on tracker import

When we import multiple trackers at once (TTM or AD import), some errors
were triggered because we checked on all fields on all trackers only for
one. This also leads to errors where some fields were mark as missing in a
tracker that does not have this field.

Change-Id: I5bab44ac0be4af294977c8e2667fe39a8d53d9c9