Introduce mermaid diagrams in trackers
Display mermaid diagrams as soon as they are enclosed in mermaid fenced
blocks (```mermaid):

* in markdown text fields
* in markdown followup comments

(To test: text field only, followup comment only, both)

The diagrams are lazy loaded: code block is replaced only when they are
in the viewport.

If no mermaid diagram detected, the lib is not included to not clutter
the assets on artifact page.

Note: only the artifact view is done. Artifact modal will come later if
needed.

Script is located in core because we will need it for git. Git is not
done directly in this patch because there is already prismjs. Mermaid
does not live well with prismjs by default. It will be done in a
dedicated commit.

Part of request #19739: Add support of mermaid js diagrams

Change-Id: I0194fa469ba04bc617287e0f3ef5f91f3edb13e5