Keep markdownish mermaid syntax in mail
<tlp-mermaid-diagram> element is only known by Tuleap web ui. In mail
notifications, we should render a regular <pre> element like other
fenced block codes so that we keep the markdownish aspect of the mermaid
syntax. (Else it is renderered as inline element and everything is on
one line… pretty difficult to read).

Part of request #19739: Add support of mermaid js diagrams

Change-Id: I8a5a611407730a52763661d613f3e640b3017ee3