Use node 15 and our own derivation of npm in the dev/build environment
Now that npm@latest targets the 7.x branch which we cannot use yet, we
cannot use the derivation provided by nixpkgs.

This contribution also solves the warnings
`The node binary used for scripts is <...> but npm is using <...> itself.`
that are currently thrown in our environment when calling a npm script.

The sub-dependency deasync is updated to avoid the need of building the
binary locally.

Test pipelines have been updated to use the same node version than one
the used in the dev environment.

Part of request #20073: Update the dev/build environment [mid-March 21]

Change-Id: I33ffa762a58ca215f0ecbb9c9f441b4ee1b8527f