Remove the 'Text' format in user preference
Part of story #18337 markdown replace text

This patch also updates from 'text' default format to 'commonmark' default format in the user preferences table in DB
Now the default user format is 'commonmark' and not 'text' anymore.

How to test:
 - Make sure to have at least one user with 'text' format preference in your DB
 - make dev-forgeupgrade
=> There is no error,
=> Users who had the 'text' format by default have now the 'commonmark'
format by default.
=> There is no change for the users who have the 'html' default format
=> In the UI of the user preferences, the 'Markdown' button radio should
be selected

 - Create a new artifact with the user who migrates from 'text' to 'commonmark' in the DB
=> All the selected format of the text fields are now 'Markdown'

You can also check this behaviours in the artifact modal or in a Step
Defintion of the artifact view.

Change-Id: I298ce11d956a08440f157f4ffe8d0927a41a9c87