Wrap syntax highlighting in a custom element
No visible changes for git (blob, tree), and artifact (text fields,
comments). The only visible change is when you edit a follow-up comment,
then the new version is syntax highlighted.

This custom element <tlp-syntax-highlighting> use IntersectionObserver
so that code blocks that never enter the viewport are not syntax
highlighted. However, while this is true for follow-up comment edition,
it is not true for the other case, since importing prismjs module
automatically trigger a highlightAll() on the page. Luckily,
highlighting seems to be idempotent.

Part of request #20075: Syntax highlights fenced code blocks in artifacts

Change-Id: Iec82510c15bcf8afaf358eabdc30932b75550ffc