Check size of committed files
When the threshold is not configured, one can commit anything.

When the threshold is set, the files can no longer be bigger
than the limit. This applies both on existing and new files.
(ie if file was bigger than the limit it can no longer be
modified).

Log errors for admins.

Part of story #19285 limit the max size of files in svn

Change-Id: Ieef41b858f329899c6af04fc8a901d7c37159fe9