Refactoring of svn PreCommit hook
As we are about to introduce another check on files
about to be commited, let's start by refactor how
the current check are done for tags and sha1 collision
detector.

Each had their own loop on files, that was not really
efficient and there were already two ways to iterate
on it.

Everything should work as before.

Part of story #19285 limit the max size of files in svn

Change-Id: Id16d1628e384ca10509c967614fbd412d567ff84