request #20090 Project add member with ldap is broken
When we add member with enabled Ldap, it will search with Tuleap Login instead of Ldap login. If they are not the same, it will not add the good user or the user will not be found

This patch will search the user with its tuleap_id before check his existance on ldap.
So, if the user is not on Tuleap, the user name sent will be the ldap one.

Change-Id: I02f54ceb5a705225c5dfd6b559360fad67f553ef