request #17195: Upgrade to Webpack 5
Some plugins have been updated to get the latest versions compatible
with Webpack 5 (and the bugfixes) and some have been exchanged for their
alternative supporting Webpack 5.

A few loaders have been removed to be replaced by the native alternative
proposed by Webpack 5.

The `exports` field of the libraries including a stylesheet have been
removed to avoid issues when importing the stylesheets. Since the
libraries are internal and we do not need the set of features proposed
by the `exports` for now it is likely we want to remove it from our
other internal libraries too in a future contribution, at the very least
to be consistent.

A deprecation remains in the MergeIntoSingleFilePlugin but it does cause
issues. A fix will be upstreamed.

Change-Id: Ie0dffa7cce3d21b1d5265bf7d89443b8b29fecd4