Enable Webpack 5 persistent caching and Babel loader cache
This should help on development environments where the same code can be
rebuilt multiple times. The persistent cache on the filesystem is not
enabled on development mode ("watch" mode in Tuleap codebase) as the in
memory cache enabled by default is likely to be a better situation.

For now the cache is not really customized. If we see too many issues
with it we can either configure it [0] or abandon it ¯\_(ツ)_/¯.

Part of request #17195: Upgrade to Webpack 5

[0] https://webpack.js.org/configuration/other-options/#cache

Change-Id: I7a7f36f1419170ed611c143334ba18891356246f