request #20093 vite: 2.0.1 -> 2.1.1 && vite-plugin-vue2: 1.2.1 -> 1.4.1
Vite and Webpack now share the same targets. This was needed because the
default target of Vite does not play with some of our code. It was
already the case before but the latest version of ESBuild (which is used
by Vite) is a bit stricter with the compatibility of some JS features
[0].

Changelogs:
https://github.com/vitejs/vite/blob/v2.1.1/packages/vite/CHANGELOG.md
https://github.com/underfin/vite-plugin-vue2/blob/9b334c90f0c7fe9683f0f51b4d1bb8b1c95eac2e/CHANGELOG.md

[0] https://github.com/evanw/esbuild/releases/tag/v0.9.3

Change-Id: I98ef13cbf43ecdef8fe10f1b865038c2ca889e4e