Remove conditional errors thrown in dev mode by the custom element relative-date
Those errors make painful any troubleshooting in one of the app using
the component as a large quantity of errors can be thrown (which will
generate a failure when troublesooting a E2E test).
It is not something that happen on non custom elements so it should
probably also not happen on custom elements.

Part of request #17195: Upgrade to Webpack 5

Change-Id: I7e3ec2374d0fada79c2f2714ab30699ade6ac392