Add permission to set build status
This is a proposition to manage the permissions for ugroups authorized
to set builds status via the route /git/:id_or_path/statuses/:commit_ref

I have chosen the ci-token pane, as this is the place to manage the
access to this route.

Also I have chosen not to use the legacy permission mecanism
using the "permission" table.

[Bonus]: I took the opportunity to remove a bit of crappiness in the
existing tab.

part of: story #18320 native support of pull requests in tuleap branch source jenkins plugin

How to test:
- Go to the ci-token pane of one of your git repositories
- Pick up several user_groups
- Submit
--> Selected user groups are marked as selected in the select box
--> Permissions are saved in db

- No functional change expected for token regeneration

Change-Id: Ia089e12b570fc84865eefd87628803b42414a9b2