The app cannot start since the upgrade to Webpack 5
The app has circular dependencies between components so the code
generated by the bundler tries to access code that is not yet defined.

We workaround the issue by using dynamic imports to hint the bundler of
the correct order.
This is however an architecture flaw of the app. The components should
not have circular dependencies between them.

Part of request #17195: Upgrade to Webpack 5

Change-Id: I2909d474ec14d76674ba60cf5707f4d9d6f563d0