Do not mangle $super argument in Prototype code
When this property is mangled Prototype is not happy about it.
Issue has been caused by the upgrade to the latest version of
terser-webpack-plugin because assets added by plugins are now
optimized [0]. This is for example the case the JS asset is
generated by the MergeIntoSingleFilePlugin plugin.

Issue was visible when creating an artifact (via the standard artifact
view) containing an Open List (e.g. a CC field).

This contribution solves multiple issues reported by the end to end
tests.

Part of request #17195: Upgrade to Webpack 5

[0] https://github.com/webpack-contrib/terser-webpack-plugin/releases/tag/v5.1.0

Change-Id: Ida109b7cdc908e2651e5e904cfd6c435149b40c0