Add / remove shortcuts on TTM controllers
Part of story #20058 : See shortcuts that are relevant for the current page in TestManagement

This patch uses the new `removeShortcutsGroup()` function from the KeyboardShortcuts lib to only display relevant shortcuts in TestManagement.
Only the shortcuts concerning the Campaigns List page are displayed on this page, and the same goes for the Campaign page.

That way, it is clearer which shortcuts are available and where.
It also allows to simplify the description of the shortcuts group by removing reference of their scope.

Change-Id: I22f24a38f0e88b1d7c715726604c1157dd0f3e72