Be able to remove a shortcuts group
Part of story #20058 : See shortcuts that are relevant for the current page in TestManagement

Makes a `removeShortcutsGroup` function available from the keyboard-shortcuts lib.
It both unbinds the shortcuts and removes the corresponding shortcuts group from the shortcuts help modal.

This patch also moves the shortcuts group elements in the shortcuts modal into dedicated <div> containers with data-attributes corresponding to the shortcuts group titles.

There should be no changes in both Document and TestManagement which have shortcuts.
The `removeShortcutsGroup` function will be used in TestManagement in
another patch.

A patch dedicated to sorting files into folder will follow, for clarification purpose.

Change-Id: I1202023c1cd4b90393849beb053f2be577ac8290