Replace the concept of CustomBlockRenderer by an extension
They both serve the same goal (extending the capabilities of the
CommonMark interpreter) so having only one concept is probably
preferable. Using the extension concept is probably to be less confusing
for a developer looking at the documentation of the league/commonmark
package.

No functional changes.

Part of request #19739: Add support of mermaid js diagrams

Change-Id: I99661e076cb8a6bf2449e7730bdfe69cd2f8246c