Inform end users about limit & default
It's hardly visible but with a bit of documentation there should
be a way for end users to know that there is a limit if they are
looking for it.

The default limit for new platform was also added.

Internal change: the limit is now defined in megabytes instead
of bytes, please adjust existing configuration.

Closes story #19285 limit the max size of files in svn

Change-Id: Ib297a9e8d63646f88c371d55a35dbb24d15e77ac