request #20118 Create or link artifact in artifact view doesn't warn user when leaving a page
Every now and then, when someone link an artifact or, worst, create an artifact
from "Artifact Link" field, they forgot to submit on the source artifact and
the link is failed.

For all other fields, as soon as one turn the edit mode, there is a warning
when the one tries to leaving the page or close the tab.
This should be consistent for artifact links.

Change-Id: I08a3c22f537416d4318452c79a7b76e065688af5