Add data-project-id on Edit comment textarea
Part of story #18337 markdown replace text

How to test:
- When you click on "Edit" on an existing follow-up comment, the
textarea that is displayed has a data-project-id attribute filled with
the current project's id.
- No other functional change

Notes:
This is mandatory for the preview button in Markdown. Since the preview
also processes Tuleap references, it needs to have a project_id passed
in the call. In the "Edit comment" case, the textarea is created by
Javascript, so it's a little bit trickier as we must first get it from
another HTML tag.

Change-Id: I23559160dc8b1d07ba55097f60e1f42daefe111d