Add the markdown "Preview" button in Artifact view
Part of story #18337 markdown replace text

How to test:

There is a "Preview" button in the following cases (in Markdown format):
- When you create a new artifact in Text fields
- When you edit an existing artifact in Text fields and new follow-up
comment
- When you edit an existing follow-up comment
- When you Mass-change artifacts and add a new follow-up comment
- When you open an existing artifact in the Modal v2 (Cardwall)

In all these cases, you can try the following:
- When you click on the "Preview" button, the format selector and the
"Help" button are disabled.
- "Preview" button becomes "Edit" button. When you click on "Edit", they
are re-enabled again.

Note: lit-html adapter is split in components to ease readability. There
is an interface before "EditorAreaState" because in the next patch it
will query a remote API. It isolates it and makes it easier to test the
renderer without mocking the remote API.
There is also another interface before TextEditor. In the next patch, we
will need to access it in the state. It leaks less implementation if we
only have access to a few methods, the State does not need to know about
CKEditor, it will only need getContent().

Change-Id: I3b94f0195f6ec12cc6eed1b279ab839961a7ba07