Everything should be accessible using a keyboard in Taskboard
Part of story #20117 : Be able to navigate with the keyboard in
Taskboard

All interactive elements should be reachable and usable using the keyboard.
Focus should be styled the same as hover, and the visibility of some elements should also be triggered by the focus.

* Visibility on focus should be enhance in a later patch.
* Taborder should also be fixed in another patch.

How to test
Given you only use the keyboard :
- You can add a card.
- You can toggle columns and swimlanes.
- You can edit the label of a card.
- You can select the assignee of a task.
- You can edit the remaining points of a story.
- You can toggle closed items (using arrows since they are radio
  inputs).

Change-Id: I7af96e184e3f69974d28d3264dd25754835bd1a5