Display months on top of Gantt
If tasks are displayed in your Roadmap widget, then months corresponding
to the chart are displayed.

If the viewport is wide enough, then extra months are added to fill the
space.

Please note that the tooltip to display the year of the month is
positionned at the bottom. This is a temporary workaround due to the
overflow hidden of the widget. It will be positionned at the top later.

Note: I hope that one day we will be able to read variables from SCSS
files into Typescript so that we don't have to repeat ourselves (here
with time unit width).

Reminder: The widget is hidden behind a feature flag:

```
tuleap config-set feature_flag_plugin_roadmap_display_underconstruction_widget 1
```

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: Ifef34de8db0fc8823c74849a51dac9573eab2802