Display today indicator
This indicator is always visible, even if today is outside from the
time period of displayed tasks.

The day of today is displayed as a native tooltip. tlp-tooltip is not
used yet since it is not easy to position it:
* widget overflow issues
* if lot of tasks, indicator may be high, the tlp-tooltip could be
displayed outside of the viewport

Technical note: A refactoring is needed so that position calculation
does not depends on "month" but on "time unit". This will be done
later in a dedicated changeset.

Reminder: The widget is hidden behind a feature flag:

```
tuleap config-set feature_flag_plugin_roadmap_display_underconstruction_widget 1
```

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: I3079e31c071b17b7db810b183e6087dd023e663f