Display task bar, nominal cases
The bar reflects the start/end dates of the corresponding task.

If no start date, or no end date, then it is a milestone.
If start date = end date, then it is a milestone.
If no start date and no end date, then the task is not displayed.
If end date < start date, then the task is not displayed.

Design note: the rotation of the div to produce a milestone may appear a
bit blurry. We will find a better way to display it (probably using svg)
in a dedicated changeset.

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: Iba45799792d4cd6b10635a0bded04d2c53455505