Display loading state
While the app is loading the tasks, a skeleton board is displayed so
that the user knows that something is going on.

Random tasks width at random position are displayed.

Reminder: The widget is hidden behind a feature flag:

```
tuleap config-set feature_flag_plugin_roadmap_display_underconstruction_widget 1
```

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: I698ca1ba4a3d61031688194da6ffb586a00c0ed6